fix: polish code for linear hash and merkle
