+++ prover/protocol/compiler/selfrecursion/column_opening.go
@@ -5,7 +5,6 @@ import (
 	"math/big"
 
 	"github.com/consensys/gnark/frontend"
-	"github.com/consensys/linea-monorepo/prover/crypto/mimc"
 	"github.com/consensys/linea-monorepo/prover/crypto/ringsis"
 	"github.com/consensys/linea-monorepo/prover/maths/common/poly"
 	"github.com/consensys/linea-monorepo/prover/maths/common/smartvectors"
@@ -14,17 +13,13 @@ import (
 	"github.com/consensys/linea-monorepo/prover/protocol/coin"
 	"github.com/consensys/linea-monorepo/prover/protocol/column"
 	"github.com/consensys/linea-monorepo/prover/protocol/column/verifiercol"
-	"github.com/consensys/linea-monorepo/prover/protocol/compiler/vortex"
 	"github.com/consensys/linea-monorepo/prover/protocol/dedicated/expr_handle"
 	"github.com/consensys/linea-monorepo/prover/protocol/dedicated/functionals"
-	"github.com/consensys/linea-monorepo/prover/protocol/dedicated/merkle"
-	mimcW "github.com/consensys/linea-monorepo/prover/protocol/dedicated/mimc"
 	"github.com/consensys/linea-monorepo/prover/protocol/ifaces"
 	"github.com/consensys/linea-monorepo/prover/protocol/wizard"
 	"github.com/consensys/linea-monorepo/prover/utils"
 	"github.com/consensys/linea-monorepo/prover/utils/gnarkutil"
 	"github.com/consensys/linea-monorepo/prover/utils/parallel"
-	"github.com/sirupsen/logrus"
 )
 
 // Specifies the column opening phase
@@ -151,349 +146,6 @@ func (ctx *SelfRecursionCtx) ColSelection() {
 			ctx.Columns.UalphaQ,
 		},
 	)
-}
-
-type linearHashMerkleProverAction struct {
-	ctx                        *SelfRecursionCtx
-	concatDhQSize              int
-	leavesSize                 int
-	leavesSizeUnpadded         int
-	sisRoundLeavesSize         int
-	sisRoundLeavesSizeUnpadded int
-	numNonSisRound             int
-	hashValuesSize             int
-	hashPreimagesSize          []int
-}
-
-func (a *linearHashMerkleProverAction) Run(run *wizard.ProverRuntime) {
-	openingIndices := run.GetRandomCoinIntegerVec(a.ctx.Coins.Q.Name)
-	concatDhQ := make([]field.Element, a.concatDhQSize*a.ctx.VortexCtx.SisParams.OutputSize())
-	merkleLeaves := make([]field.Element, a.leavesSizeUnpadded)
-	merklePositions := make([]field.Element, a.leavesSizeUnpadded)
-	merkleRoots := make([]field.Element, a.leavesSizeUnpadded)
-	// The leaves for the SIS rounds
-	sisLeaves := make([]field.Element, 0, a.sisRoundLeavesSizeUnpadded)
-	// MiMC hash metadata for non SIS rounds
-	hashValues := make([][]field.Element, 0, a.numNonSisRound)
-	hashPreimages := make([][]field.Element, 0, a.numNonSisRound)
-
-	sisHashSize := a.ctx.VortexCtx.SisParams.OutputSize()
-	numOpenedCol := a.ctx.VortexCtx.NbColsToOpen()
-	// For some reason, using a.ctx.comp.NumRounds() here does not work well here.
-	totalNumRounds := a.ctx.VortexCtx.MaxCommittedRound
-	committedRound := 0
-
-	// Handle the precomputed round
-	if a.ctx.VortexCtx.IsNonEmptyPrecomputed() {
-		rootPrecomp := a.ctx.Columns.precompRoot.GetColAssignment(run).Get(0)
-		if a.ctx.VortexCtx.IsSISAppliedToPrecomputed() {
-			precompColSisHash := a.ctx.VortexCtx.Items.Precomputeds.DhWithMerkle
-			for i, selectedCol := range openingIndices {
-				srcStart := selectedCol * sisHashSize
-				destStart := i * sisHashSize
-				sisHash := precompColSisHash[srcStart : srcStart+sisHashSize]
-				copy(concatDhQ[destStart:destStart+sisHashSize], sisHash)
-				leaf := mimc.HashVec(sisHash)
-				insertAt := i
-				merkleLeaves[insertAt] = leaf
-				sisLeaves = append(sisLeaves, leaf)
-				merkleRoots[insertAt] = rootPrecomp
-				merklePositions[insertAt].SetInt64(int64(selectedCol))
-			}
-			committedRound++
-			totalNumRounds++
-		} else {
-			precompColMiMCHash := a.ctx.VortexCtx.Items.Precomputeds.DhWithMerkle
-			precompMimcHashValues := make([]field.Element, 0, numOpenedCol)
-			precompMimcHashPreimages := make([]field.Element, 0, numOpenedCol*len(a.ctx.VortexCtx.Items.Precomputeds.PrecomputedColums))
-			for i, selectedCol := range openingIndices {
-				srcStart := selectedCol
-				// MiMC hash is a single value
-				mimcHash := precompColMiMCHash[srcStart : srcStart+1]
-				leaf := mimcHash[0]
-				mimcPreimage := a.ctx.VortexCtx.GetPrecomputedSelectedCol(selectedCol)
-				insertAt := i
-				merkleLeaves[insertAt] = leaf
-				merkleRoots[insertAt] = rootPrecomp
-				merklePositions[insertAt].SetInt64(int64(selectedCol))
-				precompMimcHashValues = append(precompMimcHashValues, leaf)
-				precompMimcHashPreimages = append(precompMimcHashPreimages, mimcPreimage...)
-			}
-			// Append the hash values and preimages
-			hashValues = append(hashValues, precompMimcHashValues)
-			hashPreimages = append(hashPreimages, precompMimcHashPreimages)
-			committedRound++
-			totalNumRounds++
-		}
-	}
-
-	for round := 0; round <= totalNumRounds; round++ {
-		if a.ctx.VortexCtx.RoundStatus[round] == vortex.IsSISApplied {
-			colSisHashName := a.ctx.VortexCtx.SisHashName(round)
-			colSisHashSV, found := run.State.TryGet(colSisHashName)
-			if !found {
-				continue
-			}
-
-			rooth := a.ctx.Columns.Rooth[round].GetColAssignment(run).Get(0)
-			colSisHash := colSisHashSV.([]field.Element)
-
-			for i, selectedCol := range openingIndices {
-				srcStart := selectedCol * sisHashSize
-				destStart := committedRound*numOpenedCol*sisHashSize + i*sisHashSize
-				sisHash := colSisHash[srcStart : srcStart+sisHashSize]
-				copy(concatDhQ[destStart:destStart+sisHashSize], sisHash)
-				leaf := mimc.HashVec(sisHash)
-				insertAt := committedRound*numOpenedCol + i
-				merkleLeaves[insertAt] = leaf
-				sisLeaves = append(sisLeaves, leaf)
-				merkleRoots[insertAt] = rooth
-				merklePositions[insertAt].SetInt64(int64(selectedCol))
-			}
-
-			run.State.TryDel(colSisHashName)
-			committedRound++
-		} else if a.ctx.VortexCtx.RoundStatus[round] == vortex.IsOnlyMiMCApplied {
-			// Fetch the MiMC hash values
-			colMimcHashName := a.ctx.VortexCtx.MIMCHashName(round)
-			colMimcHashSV, found := run.State.TryGet(colMimcHashName)
-			if !found {
-				logrus.Infof("colMimcHashName %v not found", colMimcHashName)
-				continue
-			}
-			colMimcHash := colMimcHashSV.([]field.Element)
-			// Fetch the MiMC preimages
-			nonSisOpenedColsName := a.ctx.VortexCtx.SelectedColumnNonSISName()
-			nonSisOpenedColsSV, found := run.State.TryGet(nonSisOpenedColsName)
-			if !found {
-				logrus.Infof("nonSisOpenedColsName %v not found", nonSisOpenedColsName)
-				continue
-			}
-			nonSisOpenedCols := nonSisOpenedColsSV.([][][]field.Element)
-			// Note nonSisOpenedCols contains the precomputed columns also if
-			// SIS is applied to the precomputed.
-			// However, we already have it, so we need to exclude it
-			if a.ctx.VortexCtx.IsSISAppliedToPrecomputed() {
-				nonSisOpenedCols = nonSisOpenedCols[1:]
-			}
-			// Fetch the root for the round
-			rooth := a.ctx.Columns.Rooth[round].GetColAssignment(run).Get(0)
-			mimcHashValues := make([]field.Element, 0, numOpenedCol)
-			mimcHashPreimages := make([]field.Element, 0, numOpenedCol*a.ctx.VortexCtx.GetNumPolsForNonSisRounds(round))
-			for i, selectedCol := range openingIndices {
-				srcStart := selectedCol
-				// MiMC hash is a single value
-				mimcHash := colMimcHash[srcStart : srcStart+1]
-				mimcPreimage := nonSisOpenedCols[round][i]
-				leaf := mimcHash[0]
-				insertAt := committedRound*numOpenedCol + i
-				merkleLeaves[insertAt] = leaf
-				merkleRoots[insertAt] = rooth
-				merklePositions[insertAt].SetInt64(int64(selectedCol))
-				mimcHashValues = append(mimcHashValues, leaf)
-				mimcHashPreimages = append(mimcHashPreimages, mimcPreimage...)
-			}
-			// Append the hash values and preimages
-			hashValues = append(hashValues, mimcHashValues)
-			hashPreimages = append(hashPreimages, mimcHashPreimages)
-			run.State.TryDel(colMimcHashName)
-			run.State.TryDel(nonSisOpenedColsName)
-			committedRound++
-
-		} else if a.ctx.VortexCtx.RoundStatus[round] == vortex.IsEmpty {
-			continue
-		}
-	}
-
-	numCommittedRound := a.ctx.VortexCtx.NumCommittedRounds()
-	if a.ctx.VortexCtx.IsNonEmptyPrecomputed() {
-		numCommittedRound += 1
-	}
-
-	if committedRound != numCommittedRound {
-		utils.Panic("Committed rounds %v does not match the total number of committed rounds %v", committedRound, numCommittedRound)
-	}
-
-	// Assign columns using IDs from ctx.Columns
-	run.AssignColumn(a.ctx.Columns.ConcatenatedDhQ.GetColID(), smartvectors.RightZeroPadded(concatDhQ, a.concatDhQSize))
-	run.AssignColumn(a.ctx.Columns.MerkleProofsLeaves.GetColID(), smartvectors.RightZeroPadded(merkleLeaves, a.leavesSize))
-	run.AssignColumn(a.ctx.Columns.MerkleProofPositions.GetColID(), smartvectors.RightZeroPadded(merklePositions, a.leavesSize))
-	run.AssignColumn(a.ctx.Columns.MerkleRoots.GetColID(), smartvectors.RightZeroPadded(merkleRoots, a.leavesSize))
-	run.AssignColumn(a.ctx.Columns.SisRoundLeaves.GetColID(), smartvectors.RightZeroPadded(sisLeaves, a.sisRoundLeavesSize))
-
-	// Assign the hash values and preimages
-	for i := 0; i < a.numNonSisRound; i++ {
-		run.AssignColumn(a.ctx.MIMCMetaData.ConcatenatedHashValues[i].GetColID(), smartvectors.RightZeroPadded(hashValues[i], a.hashValuesSize))
-		run.AssignColumn(a.ctx.MIMCMetaData.HashPreimages[i].GetColID(), smartvectors.RightZeroPadded(hashPreimages[i], a.hashPreimagesSize[i]))
-	}
-}
-
-func (ctx *SelfRecursionCtx) linearHashAndMerkle() {
-	roundQ := ctx.Columns.Q.Round()
-	// numRound denotes the total number of commitment rounds
-	// including SIS and non SIS rounds
-	numRound := ctx.VortexCtx.NumCommittedRounds()
-	if ctx.VortexCtx.IsNonEmptyPrecomputed() {
-		numRound += 1
-	}
-	// Next we consider the number of rounds for which we apply the SIS hash
-	numRoundSis := ctx.VortexCtx.NumCommittedRoundsSis()
-	// We increase numRoundSis by 1 if sis is applied to the precomputed
-	if ctx.VortexCtx.IsSISAppliedToPrecomputed() {
-		numRoundSis += 1
-	}
-	// The number of non SIS rounds is the difference
-	// between the total number of rounds and the number of SIS rounds
-	numRoundNonSis := numRound - numRoundSis
-
-	// The total SIS hash length = size of a single SIS hash *
-	// total number of SIS hash per SIS round * number of SIS rounds
-	concatDhQSizeUnpadded := ctx.VortexCtx.SisParams.OutputSize() * ctx.VortexCtx.NbColsToOpen() * numRoundSis
-	concatDhQSize := utils.NextPowerOfTwo(concatDhQSizeUnpadded)
-	// The leaves are computed for both SIS and non SIS rounds
-	leavesSizeUnpadded := ctx.VortexCtx.NbColsToOpen() * numRound
-	leavesSize := utils.NextPowerOfTwo(leavesSizeUnpadded)
-	// The leaves for SIS rounds
-	sisRoundLeavesSizeUnpadded := ctx.VortexCtx.NbColsToOpen() * numRoundSis
-	sisRoundLeavesSize := utils.NextPowerOfTwo(sisRoundLeavesSizeUnpadded)
-
-	ctx.Columns.ConcatenatedDhQ = ctx.comp.InsertCommit(roundQ, ctx.concatenatedDhQ(), concatDhQSize)
-	ctx.Columns.MerkleProofsLeaves = ctx.comp.InsertCommit(roundQ, ctx.merkleLeavesName(), leavesSize)
-	ctx.Columns.MerkleProofPositions = ctx.comp.InsertCommit(roundQ, ctx.merklePositionssName(), leavesSize)
-	ctx.Columns.MerkleRoots = ctx.comp.InsertCommit(roundQ, ctx.merkleRootsName(), leavesSize)
-	ctx.Columns.SisRoundLeaves = ctx.comp.InsertCommit(roundQ, ctx.sisRoundLeavesName(), sisRoundLeavesSize)
-
-	// Register the linear hash columns for non sis rounds
-	// If SIS is not applied to the precomputed, we consider
-	// it to be the first non sis round
-	ctx.MIMCMetaData.ConcatenatedHashValues = make([]ifaces.Column, 0, numRoundNonSis)
-	ctx.MIMCMetaData.HashPreimages = make([]ifaces.Column, 0, numRoundNonSis)
-	ctx.MIMCMetaData.ToHashSizes = make([]int, 0, numRoundNonSis)
-
-	// Register the linear hash columns for the non sis rounds
-	mimcHashColumnSize, mimcPreimageColumnsSize := ctx.registerMiMCMetaDataForNonSisRounds(numRoundNonSis, roundQ)
-
-	ctx.comp.RegisterProverAction(roundQ, &linearHashMerkleProverAction{
-		ctx:                ctx,
-		concatDhQSize:      concatDhQSize,
-		leavesSize:         leavesSize,
-		leavesSizeUnpadded: leavesSizeUnpadded,
-		numNonSisRound:     numRoundNonSis,
-		hashValuesSize:     mimcHashColumnSize,
-		hashPreimagesSize:  mimcPreimageColumnsSize,
-	})
-
-	depth := utils.Log2Ceil(ctx.VortexCtx.NumEncodedCols())
-
-	// The Merkle proof verification is for both sis and non sis rounds
-	merkle.MerkleProofCheck(ctx.comp, ctx.merkleProofVerificationName(), depth, leavesSizeUnpadded,
-		ctx.Columns.MerkleProofs, ctx.Columns.MerkleRoots, ctx.Columns.MerkleProofsLeaves, ctx.Columns.MerkleProofPositions)
-
-	// The linear hash verification is for only sis rounds
-	mimcW.CheckLinearHash(ctx.comp, ctx.linearHashVerificationName(), ctx.Columns.ConcatenatedDhQ,
-		ctx.VortexCtx.SisParams.OutputSize(), leavesSizeUnpadded, ctx.Columns.SisRoundLeaves)
-
-	// Register the linear hash verification for the non sis rounds
-	for i := 0; i < numRoundNonSis; i++ {
-		mimcW.CheckLinearHash(ctx.comp, ctx.nonSisRoundLinearHashVerificationName(i), ctx.MIMCMetaData.HashPreimages[i],
-			ctx.MIMCMetaData.ToHashSizes[i], ctx.VortexCtx.NbColsToOpen(), ctx.MIMCMetaData.ConcatenatedHashValues[i])
-	}
-
-	// leafConsistency imposes lookup constraints between the sis
-	// and non sis rounds leaves with that of the merkle tree leaves.
-	ctx.leafConsistency(roundQ)
-}
-
-// registerMiMCMetaDataForNonSisRounds registers the metadata for the
-// for linear hash verification for the non SIS rounds
-// and return the mimcHashColumnSize
-// and the preimage column sizes per non sis round
-func (ctx *SelfRecursionCtx) registerMiMCMetaDataForNonSisRounds(
-	numRoundNonSis int, roundQ int) (int, []int) {
-	// Compute the concatenated hashes and preimages sizes
-	var (
-		mimcHashColumnSize      = utils.NextPowerOfTwo(ctx.VortexCtx.NbColsToOpen())
-		mimcPreimageColumnsSize = make([]int, 0, numRoundNonSis)
-	)
-
-	// Consider the precomputed polynomials
-	if ctx.VortexCtx.IsNonEmptyPrecomputed() && ctx.VortexCtx.IsSISAppliedToPrecomputed() {
-		precompPreimageSize := utils.NextPowerOfTwo(
-			ctx.VortexCtx.NbColsToOpen() *
-				len(ctx.VortexCtx.Items.Precomputeds.PrecomputedColums))
-
-		ctx.MIMCMetaData.ConcatenatedHashValues = append(ctx.MIMCMetaData.ConcatenatedHashValues,
-			ctx.comp.InsertCommit(
-				roundQ,
-				ctx.concatenatedPrecomputedHashes(),
-				mimcHashColumnSize,
-			))
-
-		ctx.MIMCMetaData.HashPreimages = append(ctx.MIMCMetaData.HashPreimages,
-			ctx.comp.InsertCommit(
-				roundQ,
-				ctx.concatenatedPrecomputedPreimages(),
-				precompPreimageSize,
-			))
-		mimcPreimageColumnsSize = append(mimcPreimageColumnsSize, precompPreimageSize)
-		ctx.MIMCMetaData.ToHashSizes = append(ctx.MIMCMetaData.ToHashSizes, len(ctx.VortexCtx.Items.Precomputeds.PrecomputedColums))
-	}
-	// Next, consider only the non SIS rounds
-	for i := 0; i <= ctx.VortexCtx.NumCommittedRounds(); i++ {
-		if ctx.VortexCtx.RoundStatus[i] == vortex.IsOnlyMiMCApplied {
-
-			roundPreimageSize := utils.NextPowerOfTwo(
-				ctx.VortexCtx.NbColsToOpen() *
-					ctx.VortexCtx.GetNumPolsForNonSisRounds(i))
-
-			ctx.MIMCMetaData.ConcatenatedHashValues = append(
-				ctx.MIMCMetaData.ConcatenatedHashValues,
-				ctx.comp.InsertCommit(
-					roundQ,
-					ctx.concatenatedMiMCHashes(i),
-					mimcHashColumnSize,
-				))
-
-			ctx.MIMCMetaData.HashPreimages = append(ctx.MIMCMetaData.HashPreimages, ctx.comp.InsertCommit(
-				roundQ,
-				ctx.concatenatedMIMCPreimages(i),
-				roundPreimageSize,
-			))
-
-			ctx.MIMCMetaData.ToHashSizes = append(ctx.MIMCMetaData.ToHashSizes, ctx.VortexCtx.GetNumPolsForNonSisRounds(i))
-			mimcPreimageColumnsSize = append(mimcPreimageColumnsSize, roundPreimageSize)
-		}
-	}
-	return mimcHashColumnSize, mimcPreimageColumnsSize
-}
-
-func (ctx *SelfRecursionCtx) leafConsistency(round int) {
-	// Lookup constrains between the SIS leaves
-	// and the Merkle leaves
-	ctx.comp.InsertInclusion(
-		round,
-		ctx.sisRoundAndMerkleLeavesInclusion(),
-		[]ifaces.Column{
-			ctx.Columns.MerkleProofsLeaves,
-		},
-		[]ifaces.Column{
-			ctx.Columns.SisRoundLeaves,
-		},
-	)
-	// Lookup constrains between the non SIS leaves
-	// and the Merkle leaves
-	for i := 0; i < len(ctx.MIMCMetaData.ConcatenatedHashValues); i++ {
-		ctx.comp.InsertInclusion(
-			round,
-			ctx.nonSisRoundAndMerkleLeavesInclusion(i),
-			[]ifaces.Column{
-				ctx.Columns.MerkleProofsLeaves,
-			},
-			[]ifaces.Column{
-				ctx.MIMCMetaData.ConcatenatedHashValues[i],
-			},
-		)
-	}
 }
 
 type collapsingProverAction struct {
